construction-brain/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
faster-whisper==1.0.3
Pillow==10.4.0
python-dotenv==1.0.1
python-multipart==0.0.12
# 進度網圖 / S 曲線
plotly==5.24.1
# 文件解析
pdfplumber==0.11.4
python-docx==1.1.2
# 知識庫 RAG
chromadb==0.5.20
